+++ packages/light-client-extension-helpers/src/known-chain-specs/polkadot.ts
@@ -1,0 +1,1 @@
+export { chainSpec } from "@substrate/connect-known-chains/polkadot"