+++ packages/light-client-extension-helpers/src/known-chain-specs/rococo_v2_2.ts
@@ -1,0 +1,1 @@
+export { chainSpec } from "@substrate/connect-known-chains/rococo_v2_2"